Fix migration to use bigint for foreign key

+++ db/migrate/20190514113727_add_referent_to_stop_areas.rb
@@ -1,5 +1,5 @@
 class AddReferentToStopAreas < ActiveRecord::Migration[5.2]
   def change
-    add_column :stop_areas, :referent_id, :integer
+    add_column :stop_areas, :referent_id, :bigint
   end
 end
